Rename files to use UpperCamelCaseCluster/Logic for operational-credentials-server
