+++ src/app/clusters/operational-credentials-server/BUILD.gn
@@ -16,8 +16,8 @@ import("//build_overrides/chip.gni")
 
 source_set("operational-credentials-server") {
   sources = [
-    "operational-credentials-cluster.cpp",
-    "operational-credentials-cluster.h",
+    "OperationalCredentialsCluster.cpp",
+    "OperationalCredentialsCluster.h",
   ]
 
   public_deps = [
