+++ src/app/clusters/operational-credentials-server/OperationalCredentialsCluster.cpp
@@ -14,7 +14,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-#include <app/clusters/operational-credentials-server/operational-credentials-cluster.h>
+#include <app/clusters/operational-credentials-server/OperationalCredentialsCluster.h>
 
 #include <app/EventLogging.h>
 #include <app/InteractionModelEngine.h>
